Refactor FTL memory struct file into a new module
`FtlPrivacyLevel` is moved to the `settings` module, and `FtlQueryType`
is stored in the `counters.rs` file alongside `FtlCounters`.
