+++ src/ftl/memory_model/mod.rs
@@ -1,0 +1,22 @@
+// Pi-hole: A black hole for Internet advertisements
+// (c) 2018 Pi-hole, LLC (https://pi-hole.net)
+// Network-wide ad blocking via your own hardware.
+//
+// API
+// Shared Memory Data Types
+//
+// This file is copyright under the latest version of the EUPL.
+// Please see LICENSE file for your rights under this license.
+
+use libc;
+
+/// Used by FTL to check memory integrity in various structs
+const MAGIC_BYTE: libc::c_uchar = 0x57;
+
+mod client;
+mod counters;
+mod strings;
+
+pub use self::client::FtlClient;
+pub use self::counters::{FtlCounters, FtlQueryType};
+pub use self::strings::FtlStrings;
