+++ src/routes/stats/top_clients.rs
@@ -10,11 +10,11 @@
 
 use auth::User;
 use env::Env;
-use ftl::{FtlClient, FtlMemory, FtlPrivacyLevel};
+use ftl::{FtlClient, FtlMemory};
 use rocket::State;
 use rocket_contrib::Value;
 use routes::stats::common::{remove_excluded_clients, remove_hidden_clients};
-use settings::{ConfigEntry, FtlConfEntry};
+use settings::{ConfigEntry, FtlConfEntry, FtlPrivacyLevel};
 use util::{reply_data, Reply};
 
 /// Get the top clients with default parameters
