+++ src/settings/mod.rs
@@ -10,8 +10,10 @@
 
 mod dnsmasq;
 mod entries;
+mod privacy_level;
 mod value_type;
 
-pub use self::dnsmasq::*;
-pub use self::entries::*;
-pub use self::value_type::*;
+pub use self::dnsmasq::generate_dnsmasq_config;
+pub use self::entries::{ConfigEntry, FtlConfEntry, SetupVarsEntry};
+pub use self::privacy_level::FtlPrivacyLevel;
+pub use self::value_type::ValueType;
